Add shell.h header file
Create the shell.h header file, which contains function prototypes and includes necessary standard library headers for the Simple Shell. The header file guards (ifndef, define, endif) are used to prevent multiple inclusions of the same file. The shell.h file declares the following functions: split_line, prompt, execute, free_args, and read_line. The split_line function is responsible for splitting a line of input into individual tokens based on a specified delimiter. The prompt function displays the Shell prompt to the user. The execute function is used to execute a command with its arguments. The free_args function releases the memory allocated for the array of arguments. The read_line function reads a line from the user and returns the entered line. The shell.h header file is an essential part of the Shell program as it provides the function prototypes necessary for the Shell's functionality.