Update split_line function
Update the split_line function in the shell.c file. The function now takes an additional argument delim to specify the delimiter used for splitting the line. The rest of the function remains unchanged and continues to provide the same functionality as before, splitting the line into tokens. The update allows for more flexibility in specifying the delimiter when calling the split_line function.